backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
kiteconnect
numpy   
pandas
pandas_ta
scipy
websockets
asyncio
SQLAlchemy
gunicorn
ntplib
selenium
pyotp
webdriver-manager
psycopg2-binary
python-dotenv

# venv/bin/activate source(to be once on restart of vs code)
# gunicorn -w 1 -k uvicorn.workers.UvicornWorker main:app --bind 0.0.0.0:8000
